requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.13
django-cors-headers==4.3.1
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
msgpack==1.0.8
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
redis==5.0.6
requests==2.32.3
sqlparse==0.5.0
stripe==9.11.0
typing_extensions==4.12.2
urllib3==2.2.1
